# Optimize the Linguistic String Search Operations