This change target optimizing the string search operations (e.g. IndexOf, LastIndexOf, StartsWith, EndsWith) when using ICU. We used to create a new ICU search handle using `usearch_openFromCollator` everytime we call any of the search operations. The change here is to cache the search handles same way we cached Collation handles. The search handles are a little bit different as it needs to include the source string we need to search in and the pattern string we need to search for. This require the search handle should be used exclusively and not be shared between 2 operations in same time. Everytime we use the handle we'll set the source and the patterns strings to the handle to be able perform the operation correctly.

Caching the search handles improved the performance for IndexOf and LastIndexOf. When searching in long strings and the match pattern near the other edge of the search direction or no match found, then it shows perf improvement ranges `13% ~ 20%`. With shorter source strings the perf improvement is almost doubled. Also StartsWith/EndsWith perf is almost doubled regardless of the source string length. 


# Before #
|                      Method |       Mean |      Error |     StdDev |
|---------------------------- |-----------:|-----------:|-----------:|
|      TestMultipleStartsWith | 903.234 us | 17.9408 us | 40.4955 us |
|        TestMultipleEndsWith | 968.899 us | 19.2996 us | 38.0956 us |
|              MixLongIndexOf |  25.768 us |  0.5439 us |  1.5693 us |
|          MixLongLastIndexOf |  25.358 us |  0.5027 us |  1.3505 us |
|             MixShortIndexOf |   7.761 us |  0.1544 us |  0.3959 us |
|         MixShortLastIndexOf |   8.092 us |  0.1546 us |  0.3551 us |


# After #
|                      Method |       Mean |     Error |     StdDev | 
|---------------------------- |-----------:|----------:|-----------:|
|      TestMultipleStartsWith | 370.672 us | 7.1230 us | 17.4729 us |
|        TestMultipleEndsWith | 412.960 us | 7.9043 us | 19.2402 us |
|              MixLongIndexOf |  20.691 us | 0.4134 us |  1.1928 us |
|          MixLongLastIndexOf |  21.982 us | 0.6489 us |  1.8929 us |
|             MixShortIndexOf |   3.564 us | 0.0840 us |  0.2451 us |
|         MixShortLastIndexOf |   3.541 us | 0.0701 us |  0.1810 us |