@tarekgh 2020-10-05
https://github.com/dotnet/runtime/issues/40942

CC @adamsitnik @DrewScoggins @stephentoub