@DrewScoggins 2020-10-06
Should we bring these benchmarks into the performance suite? Seems like we should, at the very least to guard these wins for the future.