@tarekgh 2020-10-06
@DrewScoggins 

> Should we bring these benchmarks into the performance suite? Seems like we should, at the very least to guard these wins for the future.

I believe these are already covered by the tests mentioned here https://github.com/dotnet/runtime/issues/40942#issuecomment-675056206. we can look at the new numbers coming from these tests after applying this change.