@danmosemsft 2020-10-07
Just curious, does this help Windows when using ICU?