@tarekgh 2020-10-07
> Just curious, does this help Windows when using ICU?

Do you mean when directly using ICU on Windows and not from the managed code? if so, then no as the caller of the ICU is responsible for creating/opening the search handles. but it help if Windows has any code wrapping the search calls as we do in the managed.