@tarekgh 2020-10-07
@adamsitnik here is the diff results for the search perf test. Interesting the showed slower cases ar emostely ordinal and shouldn't be affected by my changes. I'll try to look more on the test itself to try make it more reliable.


| Slower                                                                           | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| -------- |
| System.Globalization.Tests.StringSearch.IsSuffix_SecondHalf(Options: (en-US, Ord |      1.33 |            16.64 |            22.15 |         |
| System.Globalization.Tests.StringSearch.IsSuffix_DifferentLastChar(Options: (en- |      1.28 |            21.03 |            26.84 |         |
| System.Globalization.Tests.StringSearch.LastIndexOf_Word_NotFound(Options: (en-U |      1.12 |           753.51 |           846.48 |         |
| System.Globalization.Tests.StringSearch.LastIndexOf_Word_NotFound(Options: (, No |      1.12 |           621.19 |           696.35 |         |
| System.Globalization.Tests.StringSearch.IsSuffix_SecondHalf(Options: (, IgnoreCa |      1.10 |           237.26 |           261.39 |         |
| System.Globalization.Tests.StringSearch.IsSuffix_DifferentLastChar(Options: (, N |      1.09 |            23.03 |            25.16 |         |
| System.Globalization.Tests.StringSearch.IsSuffix_DifferentLastChar(Options: (, I |      1.09 |          1576.73 |          1713.41 | several?|

| Faster                                                                           | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| -------- |
| System.Globalization.Tests.StringSearch.IndexOf_Word_NotFound(Options: (en-US, I |      1.96 |         16315.89 |          8344.18 | several?|
| System.Globalization.Tests.StringSearch.IndexOf_Word_NotFound(Options: (en-US, N |      1.84 |         14848.83 |          8052.22 | several?|
| System.Globalization.Tests.StringSearch.IsPrefix_FirstHalf(Options: (en-US, Ordi |      1.71 |            24.54 |            14.34 | several?|
| System.Globalization.Tests.StringSearch.IndexOf_Word_NotFound(Options: (en-US, I |      1.69 |         13450.49 |          7977.85 | several?|
| System.Globalization.Tests.StringSearch.IsSuffix_SecondHalf(Options: (en-US, Ign |      1.61 |         30757.98 |         19057.70 | bimodal |
| System.Globalization.Tests.StringSearch.IsSuffix_DifferentLastChar(Options: (en- |      1.61 |         61902.51 |         38394.12 |         |
| System.Globalization.Tests.StringSearch.IndexOf_Word_NotFound(Options: (, None,  |      1.58 |         14617.98 |          9268.79 | several?|
| System.Globalization.Tests.StringSearch.IndexOf_Word_NotFound(Options: (pl-PL, N |      1.52 |         17192.49 |         11283.35 |         |
| System.Globalization.Tests.StringSearch.LastIndexOf_Word_NotFound(Options: (en-U |      1.52 |         30131.81 |         19794.45 |         |
| System.Globalization.Tests.StringSearch.IsPrefix_FirstHalf(Options: (en-US, Igno |      1.47 |         22473.13 |         15252.63 |         |
| System.Globalization.Tests.StringSearch.IndexOf_Word_NotFound(Options: (, Ignore |      1.38 |         13604.81 |          9845.07 |         |
| System.Globalization.Tests.StringSearch.LastIndexOf_Word_NotFound(Options: (en-U |      1.37 |         25252.33 |         18385.95 |         |
| System.Globalization.Tests.StringSearch.LastIndexOf_Word_NotFound(Options: (, No |      1.37 |         25538.87 |         18641.28 |         |
| System.Globalization.Tests.StringSearch.LastIndexOf_Word_NotFound(Options: (en-U |      1.36 |         26316.24 |         19366.86 | bimodal |
| System.Globalization.Tests.StringSearch.IsSuffix_SecondHalf(Options: (en-US, Non |      1.33 |         13164.94 |          9902.41 | bimodal |
| System.Globalization.Tests.StringSearch.LastIndexOf_Word_NotFound(Options: (pl-P |      1.32 |         27810.19 |         21100.52 | several?|
| System.Globalization.Tests.StringSearch.IsPrefix_DifferentFirstChar(Options: (en |      1.31 |          1200.49 |           917.50 |         |
| System.Globalization.Tests.StringSearch.LastIndexOf_Word_NotFound(Options: (, Ig |      1.31 |         26029.60 |         19915.36 |         |
| System.Globalization.Tests.StringSearch.IsSuffix_DifferentLastChar(Options: (en- |      1.29 |          1867.69 |          1453.14 |         |
| System.Globalization.Tests.StringSearch.IndexOf_Word_NotFound(Options: (en-US, N |      1.25 |           732.24 |           585.58 | bimodal |
| System.Globalization.Tests.StringSearch.IsSuffix_SecondHalf(Options: (en-US, Non |      1.25 |           254.69 |           204.00 |         |
| System.Globalization.Tests.StringSearch.IsPrefix_DifferentFirstChar(Options: (en |      1.24 |         38068.26 |         30699.23 | several?|
| System.Globalization.Tests.StringSearch.IsSuffix_SecondHalf(Options: (, None, Tr |      1.21 |         11417.65 |          9412.09 | bimodal |
| System.Globalization.Tests.StringSearch.IsPrefix_DifferentFirstChar(Options: (en |      1.21 |            25.88 |            21.40 |         |
| System.Globalization.Tests.StringSearch.IsSuffix_DifferentLastChar(Options: (en- |      1.20 |            25.66 |            21.39 |         |
| System.Globalization.Tests.StringSearch.IsSuffix_SecondHalf(Options: (pl-PL, Non |      1.18 |         12285.80 |         10439.04 | bimodal |
| System.Globalization.Tests.StringSearch.LastIndexOf_Word_NotFound(Options: (en-U |      1.17 |           724.08 |           616.77 | bimodal |
| System.Globalization.Tests.StringSearch.IsPrefix_DifferentFirstChar(Options: (en |      1.17 |            13.57 |            11.59 | several?|
| System.Globalization.Tests.StringSearch.IsSuffix_DifferentLastChar(Options: (en- |      1.17 |          1738.64 |          1489.02 | bimodal |
| System.Globalization.Tests.StringSearch.IsPrefix_FirstHalf(Options: (en-US, None |      1.16 |          6144.75 |          5288.90 | several?|
| System.Globalization.Tests.StringSearch.IsPrefix_DifferentFirstChar(Options: (,  |      1.15 |          1079.55 |           939.51 | bimodal |
| System.Globalization.Tests.StringSearch.IsSuffix_DifferentLastChar(Options: (pl- |      1.15 |          1526.20 |          1332.11 | several?|
| System.Globalization.Tests.StringSearch.IsSuffix_DifferentLastChar(Options: (, N |      1.12 |          1767.83 |          1576.95 |         |
| System.Globalization.Tests.StringSearch.LastIndexOf_Word_NotFound(Options: (en-U |      1.11 |           127.57 |           114.49 | bimodal |
| System.Globalization.Tests.StringSearch.IsSuffix_SecondHalf(Options: (en-US, Ign |      1.11 |           247.76 |           222.57 | several?|
| System.Globalization.Tests.StringSearch.IsPrefix_DifferentFirstChar(Options: (,  |      1.11 |            24.26 |            21.85 |         |
| System.Globalization.Tests.StringSearch.IsPrefix_FirstHalf(Options: (en-US, None |      1.11 |           256.49 |           231.03 | several?|
| System.Globalization.Tests.StringSearch.IsPrefix_DifferentFirstChar(Options: (en |      1.11 |          1018.20 |           919.14 | several?|
| System.Globalization.Tests.StringSearch.IsPrefix_FirstHalf(Options: (, None, Fal |      1.10 |           232.63 |           211.51 |         |
| System.Globalization.Tests.StringSearch.IsPrefix_FirstHalf(Options: (en-US, Igno |      1.10 |           224.82 |           204.43 | several?|
| System.Globalization.Tests.StringSearch.IndexOf_Word_NotFound(Options: (en-US, I |      1.10 |           802.85 |           732.26 |         |
| System.Globalization.Tests.StringSearch.IsSuffix_SecondHalf(Options: (en-US, Ign |      1.09 |           221.54 |           203.14 |         |
| System.Globalization.Tests.StringSearch.IsPrefix_FirstHalf(Options: (, None, Tru |      1.08 |          5850.05 |          5400.00 |         |
| System.Globalization.Tests.StringSearch.LastIndexOf_Word_NotFound(Options: (en-U |      1.07 |           663.06 |           617.89 |         |
| System.Globalization.Tests.StringSearch.IsPrefix_FirstHalf(Options: (en-US, Ordi |      1.07 |            99.77 |            93.13 |         |
| System.Globalization.Tests.StringSearch.IsPrefix_FirstHalf(Options: (, IgnoreCas |      1.07 |           267.95 |           251.30 | several?|
| System.Globalization.Tests.StringSearch.IsPrefix_DifferentFirstChar(Options: (en |      1.06 |            22.72 |            21.37 |         |
| System.Globalization.Tests.StringSearch.IsPrefix_FirstHalf(Options: (en-US, Igno |      1.06 |          5558.41 |          5235.70 |         |
| System.Globalization.Tests.StringSearch.IsSuffix_DifferentLastChar(Options: (en- |      1.06 |            22.75 |            21.55 |         |
| System.Globalization.Tests.StringSearch.IsPrefix_FirstHalf(Options: (en-US, Igno |      1.05 |           252.56 |           240.76 |         |