@tarekgh 2020-10-09
@stephentoub I have addressed your feedback and now I can allow caching more search handles and that should scale well. 

@danmosemsft I have added a test for exercising the code across multiple threads to test the cache accessibility in raise conditions.

@adamsitnik @DrewScoggins I have been using the StringSearch tests and I am not noticing it is giving m different results when running it multiple times on same machine and with the exact same runtime bits (the base and the updated). I have locally changed the test a little to add a loop in every case to get more reasonable results and I think this helped and showing better results. I am pasting the results here but most the slower cases are ordinal which shouldn't be affected by my changes at all but the test showed it slower for some reason. Note, I changed the namespace, class name, and the method names just be be short and fit in the diff result.

| Slower                                                       | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| ------------------------------------------------------------ | ---------:| ----------------:| ----------------:| -------- |
| SGT.SS.IndexOf_Word_NF(Options: (en-US, Ordinal, False))     |      1.43 |         45937.12 |         65885.34 | bimodal |
| SGT.SS.LastIndexOf_Word_NF(Options: (en-US, Ordinal, False)) |      1.18 |        115238.33 |        135767.05 |         |
| SGT.SS.IsPrefix_DifferentFC(Options: (en-US, None, True))    |      1.05 |        861251.74 |        905499.82 |         |

| Faster                                                                  | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| ----------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| -------- |
| SGT.SS.IndexOf_Word_NF(Options: (, None, True))                         |      2.03 |      14149660.94 |       6985486.46 |         |
| SGT.SS.IndexOf_Word_NF(Options: (pl-PL, None, False))                   |      1.94 |      18258180.77 |       9410175.00 | bimodal |
| SGT.SS.IndexOf_Word_NF(Options: (en-US, IgnoreCase, True))              |      1.77 |      13161115.63 |       7450892.71 |         |
| SGT.SS.IndexOf_Word_NF(Options: (en-US, IgnoreSymbols, False))          |      1.76 |      12619739.06 |       7174940.63 |         |
| SGT.SS.LastIndexOf_Word_NF(Options: (pl-PL, None, False))               |      1.75 |      29810390.00 |      16993313.33 | several?|
| SGT.SS.IndexOf_Word_NF(Options: (en-US, None, True))                    |      1.71 |      12538931.25 |       7343355.21 |         |
| SGT.SS.LastIndexOf_Word_NF(Options: (en-US, IgnoreCase, True))          |      1.67 |      26948283.33 |      16151859.09 |         |
| SGT.SS.IndexOf_Word_NF(Options: (, IgnoreCase, True))                   |      1.55 |      11212507.81 |       7213828.13 |         |
| SGT.SS.IsPrefix_FH(Options: (en-US, IgnoreSymbols, False))              |      1.54 |      21852103.85 |      14145925.00 |         |
| SGT.SS.IsSuffix_DifferentLC(Options: (en-US, Ordinal, False))           |      1.47 |         27513.06 |         18709.30 |         |
| SGT.SS.IsSuffix_SH(Options: (en-US, IgnoreSymbols, False))              |      1.46 |      24717088.89 |      16901530.00 | several?|
| SGT.SS.LastIndexOf_Word_NF(Options: (en-US, None, True))                |      1.46 |      22570716.67 |      15499031.25 |         |
| SGT.SS.LastIndexOf_Word_NF(Options: (en-US, IgnoreSymbols, False))      |      1.44 |      21487418.18 |      14898540.63 |         |
| SGT.SS.LastIndexOf_Word_NF(Options: (, None, True))                     |      1.39 |      23055127.27 |      16530726.67 |         |
| SGT.SS.IsSuffix_SH(Options: (, None, True))                             |      1.39 |      10734775.00 |       7732342.19 | bimodal |
| SGT.SS.IsPrefix_DifferentFC(Options: (en-US, IgnoreSymbols, False))     |      1.37 |      36537300.00 |      26735330.00 |         |
| SGT.SS.LastIndexOf_Word_NF(Options: (, IgnoreCase, True))               |      1.36 |      21142226.92 |      15575481.25 |         |
| SGT.SS.IsPrefix_FH(Options: (pl-PL, None, False))                       |      1.27 |       7814503.13 |       6172872.92 |         |
| SGT.SS.IsSuffix_DifferentLC(Options: (en-US, OrdinalIgnoreCase, False)) |      1.26 |        189689.91 |        150495.77 | several?|
| SGT.SS.IsSuffix_DifferentLC(Options: (pl-PL, None, False))              |      1.26 |       1541743.75 |       1226774.11 |         |
| SGT.SS.IsSuffix_SH(Options: (en-US, IgnoreNonSpace, False))             |      1.25 |        235799.80 |        188080.40 |         |
| SGT.SS.IsSuffix_DifferentLC(Options: (en-US, IgnoreSymbols, False))     |      1.22 |      42056583.33 |      34420942.86 |         |
| SGT.SS.IndexOf_Word_NF(Options: (en-US, OrdinalIgnoreCase, False))      |      1.20 |        780215.63 |        652637.64 |         |
| SGT.SS.IsPrefix_DifferentFC(Options: (en-US, Ordinal, False))           |      1.18 |         10207.88 |          8632.97 |         |
| SGT.SS.IsPrefix_FH(Options: (en-US, Ordinal, False))                    |      1.18 |         14315.65 |         12122.36 |         |
| SGT.SS.IsSuffix_SH(Options: (en-US, None, True))                        |      1.18 |       9688581.25 |       8233521.88 | bimodal |
| SGT.SS.IsPrefix_DifferentFC(Options: (, None, False))                   |      1.17 |         22354.71 |         19151.02 |         |
| SGT.SS.IsPrefix_DifferentFC(Options: (en-US, IgnoreCase, True))         |      1.15 |        984311.67 |        854599.22 |         |
| SGT.SS.IsPrefix_FH(Options: (en-US, IgnoreNonSpace, False))             |      1.14 |        210044.43 |        183514.14 |         |
| SGT.SS.IsPrefix_FH(Options: (, None, True))                             |      1.14 |       5359906.25 |       4695378.91 |         |
| SGT.SS.IndexOf_Word_NF(Options: (en-US, IgnoreNonSpace, False))         |      1.12 |        601908.85 |        536391.04 | several?|
| SGT.SS.IsSuffix_DifferentLC(Options: (en-US, IgnoreCase, False))        |      1.12 |         24030.77 |         21459.00 |         |
| SGT.SS.IsPrefix_DifferentFC(Options: (, None, True))                    |      1.11 |        902719.57 |        811764.93 |         |
| SGT.SS.IsSuffix_DifferentLC(Options: (en-US, None, False))              |      1.11 |         21778.02 |         19601.82 | bimodal |
| SGT.SS.LastIndexOf_Word_NF(Options: (en-US, IgnoreCase, False))         |      1.10 |        732113.80 |        664367.75 |         |
| SGT.SS.LastIndexOf_Word_NF(Options: (en-US, OrdinalIgnoreCase, False))  |      1.10 |        780478.32 |        709557.24 |         |
| SGT.SS.IsSuffix_SH(Options: (en-US, OrdinalIgnoreCase, False))          |      1.09 |         98611.76 |         90069.54 |         |
| SGT.SS.IsPrefix_DifferentFC(Options: (en-US, OrdinalIgnoreCase, False)) |      1.08 |         11739.42 |         10824.99 | several?|
| SGT.SS.IsPrefix_FH(Options: (en-US, None, True))                        |      1.08 |       5108093.75 |       4717670.31 |         |
| SGT.SS.IsPrefix_DifferentFC(Options: (pl-PL, None, False))              |      1.08 |        879192.60 |        812896.71 |         |
| SGT.SS.IndexOf_Word_NF(Options: (, None, False))                        |      1.08 |        572535.21 |        530466.98 | bimodal |
| SGT.SS.LastIndexOf_Word_NF(Options: (en-US, IgnoreNonSpace, False))     |      1.08 |        584718.85 |        541787.28 |         |
| SGT.SS.IsPrefix_DifferentFC(Options: (en-US, None, False))              |      1.07 |         20389.91 |         19001.52 |         |
| SGT.SS.IsSuffix_SH(Options: (pl-PL, None, False))                       |      1.07 |       9538002.27 |       8912203.13 |         |
| SGT.SS.IndexOf_Word_NF(Options: (en-US, None, False))                   |      1.06 |        570469.58 |        536467.13 |         |
| SGT.SS.IsSuffix_SH(Options: (en-US, None, False))                       |      1.06 |        197056.06 |        185795.68 |         |
| SGT.SS.IsPrefix_DifferentFC(Options: (en-US, IgnoreNonSpace, False))    |      1.06 |         19985.05 |         18921.23 |         |
| SGT.SS.LastIndexOf_Word_NF(Options: (, None, False))                    |      1.05 |        570901.67 |        544106.35 |         |
| SGT.SS.LastIndexOf_Word_NF(Options: (en-US, None, False))               |      1.05 |        573550.12 |        548621.67 |         |
| SGT.SS.IsPrefix_FH(Options: (en-US, OrdinalIgnoreCase, False))          |      1.03 |         94122.27 |         91274.20 |         |